+++ app/src/main/java/com/skapps/shoppingapp/data/model/Product.kt
@@ -32,7 +32,7 @@ data class Product (
     var stockQuantity : Int?    = null,
     @ColumnInfo(name="averageRating")
     @SerializedName("averageRating" )
-    var averageRating : Int?    = null,
+    var averageRating : Double?    = null,
     @ColumnInfo(name="imageUuid")
     @SerializedName("imageUuid")
     var imageUuid : String? = null,
